Remove cyclic dependency
Issue: EF-1857
